+++ lib/widget_lifecycle.dart
@@ -1,4 +1,2 @@
-export 'src/lifecycle.dart';
-export 'src/observer.dart';
-export 'src/widget.dart';
 export 'src/state.dart';
+export 'src/widget.dart';